[clang] Allow using part of `libClangToolingCore` without depending on Clang
I'm about to make a PR that starts using `libClangToolingCore` in tblgen (to emit fixits as YAML), but since the Clang build depends on tblgen, this would create a circular dependency. To break this cycle, I have factored the parts of `libClangToolingCore` that don't depend on Clang into a separate library, `libClangToolingCoreNoClang`. To do this, I removed the Clang header includes from the "clang/tooling/Core" headers, made a few fixups to let those headers compile without the Clang types being complete, and then split the implementations in the ".cpp" files into separate files for the Clang dependencies. I then updated the CMake configuration to build the `NoClang` library first, then have the existing library depend on that.

A few random fixups were needed elsewhere in places where code was depending on a stray `using namespace llvm;` somewhere in a header that is no longer included by the tooling headers.

I also created a separate header to define a new enum, `DiagnosticLevel`, that is independent of the rest of Clang. Thus, it can be used in the tooling library or within Clang. I've updated `DiagnosticIDs::Level` to depend on this new enum. In the future we should just use this enum directly for all of the various enums that currently have separated definitions with the same values.
